+++ mobile_wallet/android/mobile_wallet_lib/src/main/java/com/concordium/mobile_wallet_lib/wallet.kt
@@ -22,6 +22,7 @@ external fun get_account_keys_and_randomness(input: String) : ReturnValue
 external fun sign_transaction(input: String) : ReturnValue
 external fun parameters_to_json(input: String) : ReturnValue
 external fun sign_message(input: String) : ReturnValue
+external fun create_unsigned_transaction(input: String) : ReturnValue
 
 fun loadWalletLib() {
     System.loadLibrary("mobile_wallet")