+++ rust-bins/wallet-notes/example.c
@@ -26,6 +26,9 @@ char* create_sec_to_pub_transfer(char*, uint8_t*);
 
 void free_response_string(char*);
 
+char* create_unsigned_transaction(char*, uint8_t*);
+char* parameters_to_json(char*, uint8_t*);
+
 /*
 $ ./example create_transfer-input.json
   calls create_transfer with the contents of create_transfer-input.json
@@ -99,6 +102,12 @@ int main(int argc, char *argv[]) {
       if (ends_with(argv[1], "create_transfer-input.json") || ends_with(argv[1], "create_transfer_with_memo-input.json")) {
         out = create_transfer(buffer, &flag);
         return printStr(out, flag);
+      } else if (ends_with(argv[1], "parameters_to_json-input.json")) {
+          out = parameters_to_json(buffer, &flag);
+          return printStr(out, flag);
+      } else if (ends_with(argv[1], "create_unsigned_transaction-input.json")) {
+          out = create_unsigned_transaction(buffer, &flag);
+          return printStr(out, flag);
       } else if (ends_with(argv[1], "create_id_request_and_private_data-input.json")) {
         out = create_id_request_and_private_data(buffer, &flag);
         return printStr(out, flag);